Language.Dockerfile

Parsing Dockerfiles (Language.Dockerfile.Syntax and Language.Dockerfile.Parser)

type Dockerfile = [InstructionPos]

Type of the Dockerfile AST

parseString :: String -> Either ParseError Dockerfile

parseFile :: String -> IO (Either ParseError Dockerfile)

Pretty-printing Dockerfiles (Language.Dockerfile.PrettyPrint)

prettyPrint :: Dockerfile -> String

Pretty print a Dockerfile to a String

prettyPrintInstructionPos :: InstructionPos -> String

Pretty print a InstructionPos to a String

Writting Dockerfiles (Language.Dockerfile.EDSL)

toDockerfileStr :: EDockerfileM a -> String

runs the Dockerfile EDSL and returns a String using PrettyPrint

import           Language.Dockerfile

main :: IO ()
main = writeFile "something.dockerfile" $ toDockerfileStr $ do
    from (tagged "fpco/stack-build" "lts-6.9")
    add "." "applanguage-dockerfile"
    workdir "applanguage-dockerfile"
    run (words "stack build --test --only-dependencies")
    cmd (words "stack test")

toDockerfile :: EDockerfileM a -> Dockerfile

Runs the Dockerfile EDSL and returns a Dockerfile you can pretty print or manipulate

toDockerfileStrIO :: MonadIO m => EDockerfileTM m t -> m String

A version of toDockerfileStr which allows IO actions

toDockerfileIO :: MonadIO m => EDockerfileTM m t -> m Dockerfile

A version of toDockerfile which allows IO actions

runDockerfileIO :: MonadIO m => EDockerfileTM m t -> m (t, Dockerfile)

Just runs the EDSL's writer monad

runDockerfileStrIO :: MonadIO m => EDockerfileTM m t -> m (t, String)

Runs the EDSL's writer monad and pretty-prints the result

liftIO :: MonadIO m => forall a. IO a -> m a

Lift a computation from the IO monad.

from :: forall m. MonadFree EInstruction m => EBaseImage -> m ()

Constructing base images

tagged :: String -> String -> EBaseImage

untagged :: String -> EBaseImage

digested :: String -> ByteString -> EBaseImage

Syntax

add :: forall m. MonadFree EInstruction m => Source -> Destination -> m ()

user :: forall m. MonadFree EInstruction m => String -> m ()

label :: forall m. MonadFree EInstruction m => Pairs -> m ()

stopSignal :: forall m. MonadFree EInstruction m => String -> m ()

copy :: forall m. MonadFree EInstruction m => Source -> Destination -> m ()

run :: MonadFree EInstruction m => String -> m ()

runArgs :: forall m. MonadFree EInstruction m => Arguments -> m ()

cmd :: MonadFree EInstruction m => String -> m ()

cmdArgs :: forall m. MonadFree EInstruction m => Arguments -> m ()

workdir :: forall m. MonadFree EInstruction m => Directory -> m ()

expose :: forall m. MonadFree EInstruction m => Ports -> m ()

ports :: [Integer] -> Ports

port :: Integer -> Ports

volume :: forall m. MonadFree EInstruction m => String -> m ()

entrypoint :: MonadFree EInstruction m => String -> m ()

entrypointArgs :: forall m. MonadFree EInstruction m => Arguments -> m ()

maintainer :: forall m. MonadFree EInstruction m => String -> m ()

env :: forall m. MonadFree EInstruction m => Pairs -> m ()

arg :: forall m. MonadFree EInstruction m => String -> m ()

comment :: forall m. MonadFree EInstruction m => String -> m ()

onBuild :: MonadFree EInstruction m => EDockerfileM a -> m ()

ONBUILD Dockerfile instruction

Each nested instruction gets emitted as a separate ONBUILD block

toDockerfile $ do
    from "node"
    run "apt-get update"
    onBuild $ do
        run "echo more-stuff"
        run "echo here"

onBuildRaw :: forall m. MonadFree EInstruction m => Instruction -> m ()

embed :: forall m. MonadFree EInstruction m => [InstructionPos] -> m ()

edockerfile :: QuasiQuoter

Quasiquoter for embedding dockerfiles on the EDSL

putStr $ toDockerfile $ do
    from "node"
    run "apt-get update"
    [edockerfile|
    RUN apt-get update
    CMD node something.js
    |]

Support types for the EDSL

type EDockerfileM = Free EInstruction

The type of Identity based EDSL blocks

type EDockerfileTM = FreeT EInstruction

The type of free monad EDSL blocks

data EBaseImage

Constructors

EUntaggedImage String

ETaggedImage String String

EDigestedImage String ByteString

QuasiQuoter (Language.Dockerfile.EDSL.Quasi)

dockerfile :: QuasiQuoter

Types (Language.Dockerfile.Syntax)

data Instruction

All commands available in Dockerfiles

Constructors

From BaseImage

Add Source Destination

User String

Label Pairs

Stopsignal String

Copy Source Destination

Run Arguments

Cmd Arguments

Workdir Directory

Expose Ports

Volume String

Entrypoint Arguments

Maintainer String

Env Pairs

Arg String

Comment String

OnBuild Instruction

EOL

data InstructionPos

Instruction with additional location information required for creating good check messages

Constructors

InstructionPos Instruction Filename Linenumber

data BaseImage

Constructors

UntaggedImage Image

TaggedImage Image Tag

DigestedImage Image ByteString

type Image = String

type Tag = String

data Ports

type Directory = String

type Source = String

type Destination = String

type Arguments = [String]

type Pairs = [(String, String)]

type Filename = String

type Linenumber = Int

Re-exports from parsec

data ParseError :: *

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.

Instruction and InstructionPos helpers

instruction :: InstructionPos -> Instruction

instructionPos :: Instruction -> InstructionPos

sourcename :: InstructionPos -> Filename