Language.Dockerfile.EDSL

Documentation

type EDockerfileM = Free EInstruction

The type of Identity based EDSL blocks

type EDockerfileTM = FreeT EInstruction

The type of free monad EDSL blocks

type EInstructionM = Free EInstruction

type EInstructionTM = FreeT EInstruction

embed :: forall m. MonadFree EInstruction m => [InstructionPos] -> m ()

onBuildRaw :: forall m. MonadFree EInstruction m => Instruction -> m ()

comment :: forall m. MonadFree EInstruction m => String -> m ()

arg :: forall m. MonadFree EInstruction m => String -> m ()

env :: forall m. MonadFree EInstruction m => Pairs -> m ()

maintainer :: forall m. MonadFree EInstruction m => String -> m ()

entrypointArgs :: forall m. MonadFree EInstruction m => Arguments -> m ()

volume :: forall m. MonadFree EInstruction m => String -> m ()

expose :: forall m. MonadFree EInstruction m => Ports -> m ()

workdir :: forall m. MonadFree EInstruction m => Directory -> m ()

cmdArgs :: forall m. MonadFree EInstruction m => Arguments -> m ()

runArgs :: forall m. MonadFree EInstruction m => Arguments -> m ()

copy :: forall m. MonadFree EInstruction m => Source -> Destination -> m ()

stopSignal :: forall m. MonadFree EInstruction m => String -> m ()

label :: forall m. MonadFree EInstruction m => Pairs -> m ()

user :: forall m. MonadFree EInstruction m => String -> m ()

add :: forall m. MonadFree EInstruction m => Source -> Destination -> m ()

from :: forall m. MonadFree EInstruction m => EBaseImage -> m ()

runDockerWriter :: MonadWriter [Instruction] m => EDockerfileM a -> m a

runDockerWriterIO :: (Monad m, MonadTrans t, Monad (t m), MonadWriter [Instruction] (t m), MonadIO (t m)) => EDockerfileTM m a -> t m a

runDef :: MonadWriter [t] m => (t1 -> t) -> t1 -> m b -> m b

runDef2 :: MonadWriter [t] m => (t1 -> t2 -> t) -> t1 -> t2 -> m b -> m b

runD :: MonadWriter [Instruction] m => EInstruction (m b) -> m b

instructionPos :: Instruction -> InstructionPos

toDockerfile :: EDockerfileM a -> Dockerfile

Runs the Dockerfile EDSL and returns a Dockerfile you can pretty print or manipulate

toDockerfileStr :: EDockerfileM a -> String

runs the Dockerfile EDSL and returns a String using PrettyPrint

import           Language.Dockerfile

main :: IO ()
main = writeFile "something.dockerfile" $ toDockerfileStr $ do
    from (tagged "fpco/stack-build" "lts-6.9")
    add "." "applanguage-dockerfile"
    workdir "applanguage-dockerfile"
    run (words "stack build --test --only-dependencies")
    cmd (words "stack test")

untagged :: String -> EBaseImage

tagged :: String -> String -> EBaseImage

digested :: String -> ByteString -> EBaseImage

ports :: [Integer] -> Ports

port :: Integer -> Ports

run :: MonadFree EInstruction m => String -> m ()

entrypoint :: MonadFree EInstruction m => String -> m ()

cmd :: MonadFree EInstruction m => String -> m ()

onBuild :: MonadFree EInstruction m => EDockerfileM a -> m ()

ONBUILD Dockerfile instruction

Each nested instruction gets emitted as a separate ONBUILD block

toDockerfile $ do
    from "node"
    run "apt-get update"
    onBuild $ do
        run "echo more-stuff"
        run "echo here"

toDockerfileIO :: MonadIO m => EDockerfileTM m t -> m Dockerfile

A version of toDockerfile which allows IO actions

toDockerfileStrIO :: MonadIO m => EDockerfileTM m t -> m String

A version of toDockerfileStr which allows IO actions

runDockerfileIO :: MonadIO m => EDockerfileTM m t -> m (t, Dockerfile)

Just runs the EDSL's writer monad

runDockerfileStrIO :: MonadIO m => EDockerfileTM m t -> m (t, String)

Runs the EDSL's writer monad and pretty-prints the result